back-end/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
dialogflow==0.7.2
distro-info===0.18ubuntu0.18.04.1
firebase-admin==3.2.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.10.2
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.16.0
idna==2.8
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
msgpack==0.6.2
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pbr==5.4.4
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.10.5
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
stevedore==1.31.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wadllib==1.3.2
xkit==0.0.0
zope.interface==4.3.2
